requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==2.0.2
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.26.100
botocore==1.29.100
bs4==0.0.1
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cleo==2.0.1
click==8.0.4
colorama==0.4.4
contourpy==1.0.6
crashtest==0.4.1
cryptography==37.0.2
cssselect==1.1.0
customtkinter==5.0.3
cycler==0.10.0
darkdetect==0.8.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.1.4
dulwich==0.21.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fake-useragent==0.1.11
fastjsonschema==2.15.3
filelock==3.9.0
fire==0.4.0
firebase-admin==6.1.0
Flask==2.0.2
fonttools==4.38.0
frozenlist==1.3.3
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.3
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
huggingface-hub==0.13.3
idna==2.10
importlib-metadata==4.12.0
install==1.3.5
installer==0.6.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
json5==0.9.8
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jwcrypto==1.3.1
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
msoffcrypto-tool==5.0.0
multidict==6.0.4
mysql-connector-python==8.0.31
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openai==0.27.3
opencv-contrib-python==4.5.5.62
opencv-python==4.7.0.72
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2022.7.4
pkginfo==1.9.6
platformdirs==2.5.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pybboxes==0.1.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
pyee==8.2.2
pyfcm==1.5.4
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.5.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==2.10.9
pyperclip==1.8.2
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyquery==1.4.3
Pyrebase4==4.5.0
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==3.3.2
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
rapidfuzz==2.13.7
requests==2.28.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roboflow==1.0.1
rsa==4.8
s3transfer==0.6.0
sahi==0.11.13
scipy==1.10.1
seaborn==0.12.2
selenium==4.3.0
Send2Trash==1.8.0
shapely==2.0.1
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
stack-data==0.3.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.15.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
torch==2.0.0
torchvision==0.15.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
trove-classifiers==2023.3.9
twilio==7.14.2
types-cryptography==3.3.23.2
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
voila==0.3.5
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.0.2
wget==3.2
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.2
yolov5==7.0.11
zipp==3.8.1
